+++ test/datalevin/search_test.clj
@@ -64,42 +64,42 @@
     (is (= (sut/doc-refs engine) [:doc1 :doc2 :doc3 :doc4 :doc5]))
 
     (let [[tid mw ^SparseIntArrayList sl]
-          (l/get-value lmdb c/terms "red" :string :term-info true)]
-      (is (= (l/range-count lmdb c/terms [:all] :string) 32))
-      (is (= (l/get-value lmdb c/terms "red" :string :int) tid))
+          (l/get-value lmdb (.-terms-dbi engine) "red" :string :term-info true)]
+      (is (= (l/range-count lmdb (.-terms-dbi engine) [:all] :string) 32))
+      (is (= (l/get-value lmdb (.-terms-dbi engine) "red" :string :int) tid))
 
       (is (sl/contains-index? sl 1))
       (is (sl/contains-index? sl 5))
       (is (= (sl/size sl) 4))
       (is (= (seq (.-indices sl)) [1 2 4 5]))
 
-      (is (= (l/list-count lmdb c/positions [tid 1] :int-int)
+      (is (= (l/list-count lmdb (.-positions-dbi engine) [tid 1] :int-int)
              (sl/get sl 1)
              2))
-      (is (= (l/list-count lmdb c/positions [tid 2] :int-int)
+      (is (= (l/list-count lmdb (.-positions-dbi engine) [tid 2] :int-int)
              (sl/get sl 2)
              1))
 
-      (is (= (l/list-count lmdb c/positions [tid 3] :int-int)
+      (is (= (l/list-count lmdb (.-positions-dbi engine) [tid 3] :int-int)
              0))
       (is (nil? (sl/get sl 3)))
 
-      (is (= (l/list-count lmdb c/positions [tid 4] :int-int)
+      (is (= (l/list-count lmdb (.-positions-dbi engine) [tid 4] :int-int)
              (sl/get sl 4)
              1))
-      (is (= (l/list-count lmdb c/positions [tid 5] :int-int)
+      (is (= (l/list-count lmdb (.-positions-dbi engine) [tid 5] :int-int)
              (sl/get sl 5)
              1))
 
-      (is (= (l/get-list lmdb c/positions [tid 5] :int-int :int-int)
+      (is (= (l/get-list lmdb (.-positions-dbi engine) [tid 5] :int-int :int-int)
              [[9 48]]))
 
-      (is (= (l/get-value lmdb c/docs 1 :int :doc-info true) [7 :doc1]))
-      (is (= (l/get-value lmdb c/docs 2 :int :doc-info true) [8 :doc2]))
-      (is (= (l/get-value lmdb c/docs 3 :int :doc-info true) [6 :doc3]))
-      (is (= (l/get-value lmdb c/docs 4 :int :doc-info true) [7 :doc4]))
-      (is (= (l/get-value lmdb c/docs 5 :int :doc-info true) [9 :doc5]))
-      (is (= (l/range-count lmdb c/docs [:all]) 5))
+      (is (= (l/get-value lmdb (.-docs-dbi engine) 1 :int :doc-info true) [7 :doc1]))
+      (is (= (l/get-value lmdb (.-docs-dbi engine) 2 :int :doc-info true) [8 :doc2]))
+      (is (= (l/get-value lmdb (.-docs-dbi engine) 3 :int :doc-info true) [6 :doc3]))
+      (is (= (l/get-value lmdb (.-docs-dbi engine) 4 :int :doc-info true) [7 :doc4]))
+      (is (= (l/get-value lmdb (.-docs-dbi engine) 5 :int :doc-info true) [9 :doc5]))
+      (is (= (l/range-count lmdb (.-docs-dbi engine) [:all]) 5))
 
       (sut/remove-doc engine :doc1)
 
@@ -107,13 +107,13 @@
       (is (= (sut/doc-refs engine) [:doc2 :doc3 :doc4 :doc5]))
 
       (let [[tid mw ^SparseIntArrayList sl]
-            (l/get-value lmdb c/terms "red" :string :term-info true)]
+            (l/get-value lmdb (.-terms-dbi engine) "red" :string :term-info true)]
         (is (not (sut/doc-indexed? engine :doc1)))
-        (is (= (l/range-count lmdb c/docs [:all]) 4))
+        (is (= (l/range-count lmdb (.-docs-dbi engine) [:all]) 4))
         (is (not (sl/contains-index? sl 1)))
         (is (= (sl/size sl) 3))
-        (is (= (l/list-count lmdb c/positions [tid 1] :int-id) 0))
-        (is (nil? (l/get-list lmdb c/positions [tid 1] :int-id :int-int)))))
+        (is (= (l/list-count lmdb (.-positions-dbi engine) [tid 1] :int-id) 0))
+        (is (nil? (l/get-list lmdb (.-positions-dbi engine) [tid 1] :int-id :int-int)))))
 
     (l/close-kv lmdb)))
 
@@ -144,6 +144,21 @@
     (is (empty? (sut/search engine "solar wind")))
     (is (= (sut/search engine "solar cap" {:display :offsets})
            [[:doc4 [["cap" [51]]]]]))
+    (l/close-kv lmdb)))
+
+(deftest multi-domains-test
+  (let [lmdb    (l/open-kv (u/tmp-dir (str "search-multi" (UUID/randomUUID))))
+        engine1 ^SearchEngine (sut/new-search-engine lmdb)
+        engine2 ^SearchEngine (sut/new-search-engine lmdb {:domain "another"})]
+    (sut/add-doc engine1 1 "hello world")
+    (sut/add-doc engine1 2 "Mars is a red planet")
+    (sut/add-doc engine1 3 "Earth is a blue planet")
+    (add-docs sut/add-doc engine2)
+
+    (is (empty? (sut/search engine1 "solar")))
+    (is (empty? (sut/search engine2 "solar")))
+    (is (= (sut/search engine1 "red") [2]))
+    (is (= (sut/search engine2 "red") [:doc1 :doc4 :doc2 :doc5]))
     (l/close-kv lmdb)))
 
 (deftest index-writer-test